Extract SearchScope and SearchInfo into Searching
